chore: sort job module related files
