+++ packages/apps/human-app/frontend/src/modules/worker/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './use-get-oracles';
