+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/components/oracles-table-item-mobile.tsx
@@ -5,9 +5,9 @@ import { TableButton } from '@/shared/components/ui/table-button';
 import { ListItem } from '@/shared/components/ui/list-item';
 import type { JobType } from '@/modules/smart-contracts/EthKVStore/config';
 import { useColorMode } from '@/shared/contexts/color-mode';
-import { type Oracle } from '../hooks';
 import { useSelectOracleNavigation } from '../hooks/use-select-oracle-navigation';
 import { EvmAddress } from '../../jobs/components';
+import { type Oracle } from '../../hooks';
 
 interface OraclesTableItemMobileProps {
   oracle: Oracle;
