+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/helpers/should-navigate-to-registration.ts
@@ -1,4 +1,4 @@
-import { type Oracle } from '../hooks';
+import { type Oracle } from '../../hooks';
 
 interface RegistrationResult {
   oracle_addresses: string[];
