+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './use-get-oracles';
 export * from './use-oracles-table-columns';
