+++ packages/apps/human-app/frontend/src/modules/worker/jobs/available-jobs/components/mobile/index.ts
@@ -1,2 +1,2 @@
+export * from './available-jobs-reward-amount-sort-mobile';
 export * from './available-jobs-table-mobile';
-export * from './available-jobs-drawer-mobile';
