+++ packages/apps/human-app/frontend/src/modules/worker/jobs/index.ts
@@ -1,0 +1,1 @@
+export * from './jobs.page';
