+++ packages/apps/human-app/frontend/src/modules/worker/jobs/jobs.page.tsx
@@ -6,18 +6,22 @@ import { useParams } from 'react-router-dom';
 import { TableQueryContextProvider } from '@/shared/components/ui/table/table-query-context';
 import { Modal } from '@/shared/components/ui/modal/modal';
 import { useIsMobile } from '@/shared/hooks/use-is-mobile';
-import { MyJobsTableMobile } from '@/modules/worker/jobs/my-jobs/mobile/my-jobs-table-mobile';
-import { AvailableJobsTable } from '@/modules/worker/jobs/available-jobs/available-jobs-table';
-import { MyJobsDrawerMobile } from '@/modules/worker/jobs/my-jobs/mobile/my-jobs-drawer-mobile';
 import { useColorMode } from '@/shared/contexts/color-mode';
 import { useGetOraclesNotifications } from '@/modules/worker/hooks/use-get-oracles-notifications';
 import { NoRecords } from '@/shared/components/ui/no-records';
-import { MyJobsTable } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-table';
 import { PageCardLoader } from '@/shared/components/ui/page-card';
-import { useGetOracles } from '../../jobs-discovery';
-import { useGetUiConfig } from '../hooks';
-import { TabPanel } from '../components';
-import { AvailableJobsDrawerMobile } from './components/mobile';
+import { useGetOracles } from '../hooks';
+import { useGetUiConfig } from './hooks';
+import { TabPanel } from './components';
+import {
+  MyJobsDrawerMobile,
+  MyJobsTableMobile,
+} from './my-jobs/components/mobile';
+import { MyJobsTable } from './my-jobs/components/desktop';
+import {
+  AvailableJobsView,
+  AvailableJobsDrawerMobileView,
+} from './available-jobs';
 
 function generateTabA11yProps(index: number) {
   return {
@@ -87,7 +91,7 @@ export function JobsPage() {
     <>
       <Modal isOpen={isMobileFilterDrawerOpen}>
         {selectedTab === 'availableJobs' && uiConfigData && (
-          <AvailableJobsDrawerMobile
+          <AvailableJobsDrawerMobileView
             chainIdsEnabled={uiConfigData.chainIdsEnabled}
             setIsMobileFilterDrawerOpen={setIsMobileFilterDrawerOpen}
           />
@@ -155,7 +159,7 @@ export function JobsPage() {
                     {isError ? (
                       <NoRecords />
                     ) : (
-                      <AvailableJobsTable
+                      <AvailableJobsView
                         chainIdsEnabled={uiConfigData.chainIdsEnabled}
                         handleOpenMobileFilterDrawer={
                           handleOpenMobileFilterDrawer
