+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/desktop/index.ts
@@ -1,0 +1,1 @@
+export * from './my-jobs-table';
