+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/desktop/my-jobs-expires-at-sort.tsx
@@ -1,23 +1,23 @@
 /* eslint-disable camelcase --- ... */
 import { t } from 'i18next';
 import { Sorting } from '@/shared/components/ui/table/table-header-menu.tsx/sorting';
-import { useMyJobsFilterStore } from '../../hooks';
+import { useMyJobsFilterStore } from '../../../hooks';
 
-export function MyJobsRewardAmountSort() {
+export function MyJobsExpiresAtSort() {
   const { setFilterParams, filterParams } = useMyJobsFilterStore();
 
-  const sortAscRewardAmount = () => {
+  const sortAscExpiresAt = () => {
     setFilterParams({
       ...filterParams,
-      sort_field: 'reward_amount',
+      sort_field: 'expires_at',
       sort: 'asc',
     });
   };
 
-  const sortDescRewardAmount = () => {
+  const sortDescExpiresAt = () => {
     setFilterParams({
       ...filterParams,
-      sort_field: 'reward_amount',
+      sort_field: 'expires_at',
       sort: 'desc',
     });
   };
@@ -33,12 +33,12 @@ export function MyJobsRewardAmountSort() {
       }}
       sortingOptions={[
         {
-          label: t('worker.jobs.sortDirection.fromHighest'),
-          sortCallback: sortDescRewardAmount,
+          label: t('worker.jobs.sortDirection.closestToNow'),
+          sortCallback: sortAscExpiresAt,
         },
         {
-          label: t('worker.jobs.sortDirection.fromLowest'),
-          sortCallback: sortAscRewardAmount,
+          label: t('worker.jobs.sortDirection.furthestToNow'),
+          sortCallback: sortDescExpiresAt,
         },
       ]}
     />
