+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/desktop/my-jobs-table.tsx
@@ -13,26 +13,26 @@ import { TableHeaderCell } from '@/shared/components/ui/table/table-header-cell'
 import { getNetworkName } from '@/modules/smart-contracts/get-network-name';
 import { Button } from '@/shared/components/ui/button';
 import { Chip } from '@/shared/components/ui/chip';
-import { MyJobsJobTypeFilter } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-job-type-filter';
-import { MyJobsRewardAmountSort } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-reward-amount-sort';
-import { MyJobsStatusFilter } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-status-filter';
-import { MyJobsExpiresAtSort } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-expires-at-sort';
-import { MyJobsNetworkFilter } from '@/modules/worker/jobs/my-jobs/desktop/my-jobs-network-filter';
 import { useColorMode } from '@/shared/contexts/color-mode';
 import { createTableDarkMode } from '@/shared/styles/create-table-dark-mode';
 import type { JobType } from '@/modules/smart-contracts/EthKVStore/config';
 import { formatDate } from '@/shared/helpers/date';
-import { useMyJobsFilterStore } from '../../hooks';
 import { StatusChip } from '../status-chip';
-import { useRefreshTasksMutation } from '../hooks';
 import {
-  EscrowAddressSearchForm,
-  EvmAddress,
   type MyJob,
-  MyJobsTableActions,
+  EvmAddress,
   RewardAmount,
+  MyJobsTableActions,
   useGetMyJobsData,
-} from '../../components';
+  EscrowAddressSearchForm,
+} from '../../../components';
+import { useMyJobsFilterStore } from '../../../hooks';
+import { useRefreshTasksMutation } from '../../hooks';
+import { MyJobsExpiresAtSort } from './my-jobs-expires-at-sort';
+import { MyJobsJobTypeFilter } from './my-jobs-job-type-filter';
+import { MyJobsNetworkFilter } from './my-jobs-network-filter';
+import { MyJobsRewardAmountSort } from './my-jobs-reward-amount-sort';
+import { MyJobsStatusFilter } from './my-jobs-status-filter';
 
 interface MyJobsTableProps {
   chainIdsEnabled: number[];
