+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/mobile/index.ts
@@ -1,0 +1,2 @@
+export * from './my-jobs-drawer-mobile';
+export * from './my-jobs-table-mobile';
