+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/mobile/my-jobs-expires-at-sort-mobile.tsx
@@ -2,10 +2,10 @@
 import Typography from '@mui/material/Typography';
 import { t } from 'i18next';
 import { useColorMode } from '@/shared/contexts/color-mode';
-import { useMyJobsFilterStore } from '../../hooks';
-import { Sorting } from '../../components';
+import { Sorting } from '../../../components';
+import { useMyJobsFilterStore } from '../../../hooks';
 
-export function MyJobsRewardAmountSortMobile() {
+export function MyJobsExpiresAtSortMobile() {
   const { setFilterParams, filterParams } = useMyJobsFilterStore();
   const { colorPalette } = useColorMode();
 
@@ -13,29 +13,27 @@ export function MyJobsRewardAmountSortMobile() {
     <Sorting
       label={
         <Typography color={colorPalette.text.secondary} variant="body2">
-          {t('worker.jobs.rewardAmount')}
+          {t('worker.jobs.expiresAt')}
         </Typography>
       }
       fromHighestSelected={
-        filterParams.sort_field === 'reward_amount' &&
-        filterParams.sort === 'desc'
+        filterParams.sort_field === 'expires_at' && filterParams.sort === 'desc'
       }
       sortFromHighest={() => {
         setFilterParams({
           ...filterParams,
           sort: 'desc',
-          sort_field: 'reward_amount',
+          sort_field: 'expires_at',
         });
       }}
       fromLowestSelected={
-        filterParams.sort_field === 'reward_amount' &&
-        filterParams.sort === 'asc'
+        filterParams.sort_field === 'expires_at' && filterParams.sort === 'asc'
       }
       sortFromLowest={() => {
         setFilterParams({
           ...filterParams,
           sort: 'asc',
-          sort_field: 'reward_amount',
+          sort_field: 'expires_at',
         });
       }}
       clear={() => {
