+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/mobile/my-jobs-job-type-filter-mobile.tsx
@@ -2,7 +2,7 @@
 import { useTranslation } from 'react-i18next';
 import { Filtering } from '@/shared/components/ui/table/table-header-menu.tsx/filtering';
 import { JOB_TYPES } from '@/shared/consts';
-import { useMyJobsFilterStore } from '../../hooks';
+import { useMyJobsFilterStore } from '../../../hooks';
 
 export function MyJobsJobTypeFilterMobile() {
   const { t } = useTranslation();
