+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/mobile/my-jobs-reward-amount-sort-mobile.tsx
@@ -2,10 +2,10 @@
 import Typography from '@mui/material/Typography';
 import { t } from 'i18next';
 import { useColorMode } from '@/shared/contexts/color-mode';
-import { useMyJobsFilterStore } from '../../hooks';
-import { Sorting } from '../../components';
+import { Sorting } from '../../../components';
+import { useMyJobsFilterStore } from '../../../hooks';
 
-export function MyJobsExpiresAtSortMobile() {
+export function MyJobsRewardAmountSortMobile() {
   const { setFilterParams, filterParams } = useMyJobsFilterStore();
   const { colorPalette } = useColorMode();
 
@@ -13,27 +13,29 @@ export function MyJobsExpiresAtSortMobile() {
     <Sorting
       label={
         <Typography color={colorPalette.text.secondary} variant="body2">
-          {t('worker.jobs.expiresAt')}
+          {t('worker.jobs.rewardAmount')}
         </Typography>
       }
       fromHighestSelected={
-        filterParams.sort_field === 'expires_at' && filterParams.sort === 'desc'
+        filterParams.sort_field === 'reward_amount' &&
+        filterParams.sort === 'desc'
       }
       sortFromHighest={() => {
         setFilterParams({
           ...filterParams,
           sort: 'desc',
-          sort_field: 'expires_at',
+          sort_field: 'reward_amount',
         });
       }}
       fromLowestSelected={
-        filterParams.sort_field === 'expires_at' && filterParams.sort === 'asc'
+        filterParams.sort_field === 'reward_amount' &&
+        filterParams.sort === 'asc'
       }
       sortFromLowest={() => {
         setFilterParams({
           ...filterParams,
           sort: 'asc',
-          sort_field: 'expires_at',
+          sort_field: 'reward_amount',
         });
       }}
       clear={() => {
