+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/components/mobile/my-jobs-status-filter-mobile.tsx
@@ -1,9 +1,9 @@
 import capitalize from 'lodash/capitalize';
 import { Filtering } from '@/shared/components/ui/table/table-header-menu.tsx/filtering';
-import { useMyJobsFilterStore } from '../../hooks';
-import { MyJobStatus } from '../../types';
+import { useMyJobsFilterStore } from '../../../hooks';
+import { MyJobStatus } from '../../../types';
 
-export function MyJobsStatusFilter() {
+export function MyJobsStatusFilterMobile() {
   const { setFilterParams, filterParams } = useMyJobsFilterStore();
 
   return (
@@ -12,6 +12,7 @@ export function MyJobsStatusFilter() {
         setFilterParams({
           ...filterParams,
           status: undefined,
+          page: 0,
         });
       }}
       filteringOptions={Object.values(MyJobStatus).map((status) => ({
@@ -23,6 +24,7 @@ export function MyJobsStatusFilter() {
         setFilterParams({
           ...filterParams,
           status,
+          page: 0,
         });
       }}
     />
